requirements.txt
Django==5.1
djangorestframework==3.15.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dotenv==1.0.1
PyJWT==2.9.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
celery==5.4.0
redis==5.0.8
drf-yasg==1.21.7
setuptools==72.1.0
pillow==10.4.0
coverage==7.6.1

